+++ recipes/kube-controller.rb
@@ -5,12 +5,16 @@
 # Copyright 2015-2016, Bloomberg Finance L.P.
 #
 
-service 'kube-controller-manager' do
-  action :enable
+file '/var/log/kube-controller-manager.log' do
+  action :touch
 end
 
-template '/etc/kubernetes/controller-manager' do
+template '/etc/kubernetes/inactive-manifests/controller-manager.yaml' do
   mode '0640'
   source 'kube-controller-manager.erb'
-  notifies :restart, 'service[kube-controller-manager]', :immediately
+  variables(
+    controller_manager_image: node['kubernetes']['master']['controller-manager-source'],
+    kubernetes_api_port: node['kubernetes']['insecure']['apiport'],
+    etcd_cert_dir: node['kubernetes']['secure']['directory']
+  )
 end