+++ recipes/kube-scheduler.rb
@@ -5,12 +5,16 @@
 # Copyright 2015-2016, Bloomberg Finance L.P.
 #
 
-service 'kube-scheduler' do
-  action :enable
+file '/var/log/kube-scheduler.log' do
+  action :touch
 end
 
-template '/etc/kubernetes/scheduler' do
+template '/etc/kubernetes/inactive-manifests/scheduler.yaml' do
   mode '0640'
   source 'kube-scheduler.erb'
-  notifies :restart, 'service[kube-scheduler]', :immediately
+  variables(
+    kube_scheduler_image: node['kubernetes']['master']['scheduler-source'],
+    kubernetes_api_port: node['kubernetes']['insecure']['apiport'],
+    etcd_cert_dir: node['kubernetes']['secure']['directory']
+  )
 end