+++ recipes/podmaster.rb
@@ -5,14 +5,12 @@
 # Copyright 2015-2016, Bloomberg Finance L.P.
 #
 
-template '/etc/kubernetes/config' do
+template '/etc/kubernetes/manifests/podmaster.yaml' do
   mode '0640'
-  source 'kube-config.erb'
+  source 'podmaster.yaml.erb'
   variables(
-    kubernetes_log_level: node['kubernetes']['log']['level'],
-    kubernetes_master: node['fqdn'],
+    podmaster_image: node['kubernetes']['master']['podmaster-source'],
     kubernetes_api_port: node['kubernetes']['insecure']['apiport'],
-    kubernetes_secure_api_port: node['kubernetes']['secure']['apiport'],
     etcd_cert_dir: node['kubernetes']['secure']['directory']
   )
 end